requirements.txt
# SDKs
git+https://github.com/WIPtitle/rabbitmq-sdk.git

# Other deps
fastapi~=0.109.2
uvicorn==0.30.6
psycopg2==2.9.9
sqlmodel==0.0.22
httpx==0.26.0
pillow==10.4.0
requests==2.32.3